+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Calling `Model::paginate()` will no longer raise an error.
+
 ## [0.1.1] 2020-09-10
 
 ### Fixed